refactoring frontend
