+++ frontend/src/features/request/index.js
@@ -1,19 +1,18 @@
 // Main component
-export { default as RequestDetails } from "./RequestDetails";
+export { default as RequestDetails } from "./components/RequestDetails";
 
 // Sub-components
-export { default as EmptyState } from "./EmptyState";
-export { default as RequestCard } from "./RequestCard";
-export { default as RequestHeader } from "./RequestHeader";
-export { default as FormsSection } from "./FormsSection";
-export { default as NotesSection } from "./NotesSection";
-export { default as ActionButtons } from "./ActionButtons";
-export { default as FormInfo } from "./FormInfo";
+export { default as EmptyState } from "./components/EmptyState";
+export { default as RequestCard } from "./components/RequestCard";
+export { default as RequestHeader } from "./components/RequestHeader";
+export { default as FormsSection } from "./components/FormsSection";
+export { default as NotesSection } from "./components/NotesSection";
+export { default as FormInfo } from "./components/FormInfo";
 
 // Existing components (if you want to re-export them)
-export { default as RequestedDoc } from "./RequestedDoc";
-export { default as RequestedDocsList } from "./RequestedDocsList";
-export { default as RequestRow } from "./RequestRow";
-export { default as RequestsTable } from "./RequestsTable";
-export { default as RequestsTableOperations } from "./RequestsTableOperations";
-export { default as NewRequestForm } from "./NewRequestForm";
+export { default as RequestedDoc } from "./components/RequestedDoc";
+export { default as RequestedDocsList } from "./components/RequestedDocsList";
+export { default as RequestRow } from "./components/RequestRow";
+export { default as RequestsTable } from "./components/RequestsTable";
+export { default as RequestsTableOperations } from "./components/RequestsTableOperations";
+export { default as NewRequestForm } from "./components/NewRequestForm";
